# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: CRE Loan Origination System API
servers:
  - url: /java-api
    description: Spring Boot API
...
paths:
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all loan applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: phase
          in: query
          schema:
            type: string
            enum: [Inquiry, ICR, Application, FCR, Closing, Disposed]
        - name: search
          in: query
          schema:
            type: string
        - name: rateType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Application list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanApplication"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new loan application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get loan application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplicationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateApplication
      tags: [applications]
      summary: Update a loan application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationRequest"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete a loan application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /applications/{id}/advance-status:
    post:
      operationId: advanceApplicationStatus
      tags: [applications]
      summary: Advance application to next workflow status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvanceStatusRequest"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"

  # ── Loan Terms ─────────────────────────────────────────────────────────────
  /applications/{id}/loan-terms:
    get:
      operationId: getLoanTerms
      tags: [loanTerms]
      summary: Get loan terms for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Loan terms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanTerms"
    put:
      operationId: updateLoanTerms
      tags: [loanTerms]
      summary: Update loan terms for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanTerms"
      responses:
        "200":
          description: Updated loan terms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanTerms"

  # ── Comments ───────────────────────────────────────────────────────────────
  /applications/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List threaded comments for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /applications/{id}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Documents ──────────────────────────────────────────────────────────────
  /applications/{id}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Add a document record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /applications/{id}/documents/{docId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: docId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Borrowers ──────────────────────────────────────────────────────────────
  /borrowers:
    get:
      operationId: listBorrowers
      tags: [borrowers]
      summary: List all borrowers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Borrower list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Borrower"
    post:
      operationId: createBorrower
      tags: [borrowers]
      summary: Create a borrower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBorrowerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrower"
  /borrowers/{id}:
    get:
      operationId: getBorrower
      tags: [borrowers]
      summary: Get borrower by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Borrower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrower"
    put:
      operationId: updateBorrower
      tags: [borrowers]
      summary: Update a borrower
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBorrowerRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrower"

  # ── Properties ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ActivityItem:
      type: object
      required: [id, applicationId, action, description, timestamp]
      properties:
        id:
          type: string
        applicationId:
          type: string
        action:
          type: string
        description:
          type: string
        userName:
          type: string
        timestamp:
          type: string
          format: date-time
...
    LoanApplication:
      type: object
      required: [id, status, borrowerId, propertyId, loanType, loanAmountUsd, createdAt, updatedAt]
      properties:
        id:
          type: string
        status:
          type: string
        borrowerId:
          type: string
        propertyId:
          type: string
        loanType:
          type: string
        loanAmountUsd:
          type: string
        loanTermYears:
          type: string
        interestType:
          type: string
        interestRatePct:
          type: string
        amortizationType:
          type: string
        ltvPct:
          type: string
        dscrRatio:
          type: string
        targetClosingDate:
          type: string
        rateType:
          type: string
        allInFixedRate:
          type: string
        proformaAdjustableAllInRate:
          type: string
        borrowerName:
          type: string
        entityName:
          type: string
        propertyAddress:
          type: string
        propertyCity:
          type: string
        propertyState:
          type: string
        propertyType:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LoanApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/LoanApplication"
        - type: object
          properties:
            borrower:
              $ref: "#/components/schemas/Borrower"
            property:
              $ref: "#/components/schemas/Property"
            loanTerms:
              $ref: "#/components/schemas/LoanTerms"
            commentCount:
              type: integer
            documentCount:
              type: integer
    CreateApplicationRequest:
      type: object
      required: [borrowerId, propertyId, loanType, loanAmountUsd]
      properties:
        borrowerId:
          type: string
        propertyId:
          type: string
        loanType:
          type: string
        loanAmountUsd:
          type: string
        loanTermYears:
          type: string
        interestType:
          type: string
        amortizationType:
          type: string
        ltvPct:
          type: string
        dscrRatio:
          type: string
        targetClosingDate:
          type: string
    UpdateApplicationRequest:
      type: object
      properties:
        status:
          type: string
        loanType:
          type: string
        loanAmountUsd:
          type: string
        loanTermYears:
          type: string
        interestType:
          type: string
        interestRatePct:
          type: string
        amortizationType:
          type: string
        ltvPct:
          type: string
        dscrRatio:
          type: string
        targetClosingDate:
          type: string
    AdvanceStatusRequest:
      type: object
      required: [nextStatus]
      properties:
        nextStatus:
          type: string

    # ── LoanTerms ─────────────────────────────────────────────────────────────
    LoanTerms:
      type: object
      properties:
        applicationId:
          type: string
        rateType:
          type: string
        baseRate:
          type: string
        fixedRateVariance:
          type: string
        indexName:
          type: string
        indexRate:
          type: string
        spreadOnFixed:
          type: string
        allInFixedRate:
          type: string
        adjustableRateVariance:
          type: string
        adjustableIndexName:
          type: string
        adjustableIndexRate:
          type: string
        spreadOnAdjustable:
          type: string
        proformaAdjustableAllInRate:
          type: string
        loanAmountUsd:
          type: string
        loanTermYears:
          type: string
        amortizationYears:
          type: string
        amortizationType:
          type: string
        ltvPct:
          type: string
        dscrRatio:
          type: string
        targetClosingDate:
          type: string
        interestOnlyPeriodMonths:
          type: string
        prepaymentPenaltyPct:
          type: string
        originationFeePct:
          type: string
        ecoaNoticeDate:
          type: string
        ecoaActionTaken:
          type: string

    # ── Comments ──────────────────────────────────────────────────────────────
    Comment:
      type: object
      required: [id, applicationId, body, authorName, createdAt]
      properties:
        id:
          type: string
        applicationId:
          type: string
        parentCommentId:
          type: string
          nullable: true
        body:
          type: string
        authorName:
          type: string
        authorSid:
          type: string
        createdAt:
          type: string
          format: date-time
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    CreateCommentRequest:
      type: object
      required: [body, authorName]
      properties:
        body:
          type: string
        authorName:
          type: string
        authorSid:
          type: string
        parentCommentId:
          type: string

    # ── Documents ─────────────────────────────────────────────────────────────
    Document:
      type: object
      required: [id, applicationId, fileName, fileType, uploadedBy, uploadedAt]
      properties:
        id:
          type: string
        applicationId:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        fileSizeBytes:
          type: integer
        category:
          type: string
        uploadedBy:
          type: string
        uploadedAt:
          type: string
          format: date-time
        description:
          type: string
    CreateDocumentRequest:
      type: object
      required: [fileName, fileType, uploadedBy]
      properties:
        fileName:
          type: string
        fileType:
          type: string
        fileSizeBytes:
          type: integer
        category:
          type: string
        uploadedBy:
          type: string
        description:
          type: string

    # ── Borrower ──────────────────────────────────────────────────────────────
    Borrower:
      type: object
      required: [id, firstName, lastName, createdAt, updatedAt]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        entityName:
          type: string
        email:
          type: string
        phone:
          type: string
        mailingAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        creExperienceYears:
          type: string
        netWorthUsd:
          type: string
        liquidityUsd:
          type: string
        creditScore:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBorrowerRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        entityName:
          type: string
        email:
          type: string
        phone:
          type: string
        mailingAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        creExperienceYears:
          type: string
        netWorthUsd:
          type: string
        liquidityUsd:
          type: string
        creditScore:
          type: string

    # ── Property ──────────────────────────────────────────────────────────────
    Property:
      type: object
      required: [id, streetAddress, city, state, zipCode, propertyType, createdAt, updatedAt]
      properties:
        id:
          type: string
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        propertyType:
          type: string
        grossSqFt:
          type: string
        numberOfUnits:
          type: string
        yearBuilt:
          type: string
        physicalOccupancyPct:
          type: string
        economicOccupancyPct:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiError:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string