# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRE Loan Origination System API
 * OpenAPI spec version: 1.0.0
 */

export interface ActivityItem {
  id: string;
  applicationId: string;
  action: string;
  description: string;
  userName?: string;
  timestamp: Date;
}