# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRE Loan Origination System API
 * OpenAPI spec version: 1.0.0
 */

export interface ApiError {
  message: string;
  code?: string;
}